Move queries to separate file
